chore(dependency): bump all dependencies to latest and upgrade to Nextjs App Dir
