+++ package.json
@@ -13,30 +13,30 @@
     "start:static": "npx http-server ./out"
   },
   "dependencies": {
-    "clsx": "^1.2.1",
-    "next": "13.4.4",
+    "clsx": "^2.0.0",
+    "next": "14.0.2",
     "react": "latest",
     "react-dom": "latest",
-    "swiper": "^9.3.2",
-    "swr": "^2.1.5"
+    "swiper": "^11.0.4",
+    "swr": "^2.2.4"
   },
   "devDependencies": {
-    "@types/node": "18.15.11",
-    "@types/react": "18.0.33",
-    "@typescript-eslint/eslint-plugin": "^5.59.7",
-    "@typescript-eslint/parser": "^5.59.7",
-    "autoprefixer": "^10.4.14",
-    "eslint": "^8.41.0",
-    "eslint-config-prettier": "^8.8.0",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react": "^7.32.2",
+    "@types/node": "20.9.0",
+    "@types/react": "18.2.37",
+    "@typescript-eslint/eslint-plugin": "^6.10.0",
+    "@typescript-eslint/parser": "^6.10.0",
+    "autoprefixer": "^10.4.16",
+    "eslint": "^8.53.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.1.2",
-    "prettier": "2.8.8",
-    "react-icons": "^4.8.0",
-    "sass": "^1.62.1",
-    "tailwindcss": "^3.3.2",
-    "typescript": "^5.0.4"
+    "prettier": "3.0.3",
+    "react-icons": "^4.11.0",
+    "sass": "^1.69.5",
+    "tailwindcss": "^3.3.5",
+    "typescript": "^5.2.2"
   },
   "repository": {
     "type": "git",
