+++ src/components/InvisibleLabel.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { useEffect, useRef, useState } from "react"
 import clsx from "clsx"
