--- src/pages/404.tsx
@@ -1,10 +1,0 @@
-import * as React from "react"
-
-const Error404Page = () => (
-  <main>
-    <h1>啊！你似乎来错了地方</h1>
-    <h2>请返回再看看别的吧。</h2>
-  </main>
-)
-
-export default Error404Page
