+++ tsconfig.json
@@ -1,17 +1,29 @@
 {
   "include": [
-    "./src/**/*.tsx"
+    "./src/**/*.tsx",
+    ".next/types/**/*.ts"
   ],
-  
   "compilerOptions": {
     "baseUrl": ".",
     "paths": {
-      "@/components/*": ["src/components/*"],
-      "@/pages/*": ["src/pages/*"],
-      "@/styles/*": ["src/styles/*"],
-      "@/utils/*": ["src/utils/*"],
-      "@/content/*": ["content/*"],
-      "@/types/*": ["src/types/*"],
+      "@/components/*": [
+        "src/components/*"
+      ],
+      "@/app/*": [
+        "src/app/*"
+      ],
+      "@/styles/*": [
+        "src/styles/*"
+      ],
+      "@/utils/*": [
+        "src/utils/*"
+      ],
+      "@/content/*": [
+        "content/*"
+      ],
+      "@/types/*": [
+        "src/types/*"
+      ],
     },
     "target": "esnext",
     "module": "esnext",
@@ -31,7 +43,12 @@
     "allowJs": true,
     "forceConsistentCasingInFileNames": true,
     "incremental": true,
-    "isolatedModules": true
+    "isolatedModules": true,
+    "plugins": [
+      {
+        "name": "next"
+      }
+    ]
   },
   "exclude": [
     "node_modules"
